Altered cooking dialog to show all ingredients at once, and changed selection style.
The previous dialog only showed three ingredients at a time, a style that didn't lend itself to
contemplation of the recipe puzzles. I also felt that changing the border color wasn't enough of
an indication of selection, so now I use a "bloom" effect obtained by using a text shadow the
same color as the text.