Implements new configurable scene model, and explicit dialogue handling in the HUD.
The use of a scene "goal" was causing loads of duplication in the scene staging. What I'd like
to try is simply having all the parameters of the scene in the lookup--is there a speech, are
the controls locked out, what's the NPC doing--rather than having that functionality assigned to
a label.

Also, as I'm adopting more of a dialogue approach to the script, I want to establish this.